refactor package and test
