+++ examples/camunda-spring-boot-starter-example-simple/src/main/java/org/camunda/bpm/spring/boot/starter/example/simple/Showcase.java
@@ -5,7 +5,7 @@ import static org.slf4j.LoggerFactory.getLogger;
 import org.camunda.bpm.engine.RuntimeService;
 import org.camunda.bpm.engine.TaskService;
 import org.camunda.bpm.engine.task.Task;
-import org.camunda.bpm.spring.boot.starter.events.ProcessApplicationStartedEvent;
+import org.camunda.bpm.spring.boot.starter.event.ProcessApplicationStartedEvent;
 import org.slf4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.event.EventListener;
