+++ extension/camunda-spring-boot-starter/src/main/java/org/camunda/bpm/spring/boot/starter/event/JobExecutorStartingEventListener.java
@@ -1,4 +1,4 @@
-package org.camunda.bpm.spring.boot.starter.events;
+package org.camunda.bpm.spring.boot.starter.event;
 
 import org.camunda.bpm.engine.impl.jobexecutor.JobExecutor;
 import org.slf4j.Logger;
