+++ extension/camunda-spring-boot-starter/src/test/java/org/camunda/bpm/spring/boot/starter/event/ProcessApplicationEventPublisherTest.java
@@ -1,6 +1,7 @@
-package org.camunda.bpm.spring.boot.starter.events;
+package org.camunda.bpm.spring.boot.starter.event;
 
 import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
@@ -8,6 +9,7 @@ import static org.mockito.Mockito.when;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.runners.MockitoJUnitRunner;
@@ -19,6 +21,10 @@ import org.springframework.context.event.ContextClosedEvent;
 @RunWith(MockitoJUnitRunner.class)
 public class ProcessApplicationEventPublisherTest {
 
+
+  @InjectMocks
+  private ProcessApplicationEventPublisher processApplicationEventPublisher;
+
   @Mock
   private ApplicationEventPublisher publisherMock;
 
@@ -27,23 +33,22 @@ public class ProcessApplicationEventPublisherTest {
 
   @Test
   public void handleApplicationReadyEventTest() {
-    ProcessApplicationEventPublisher processApplicationEventPublisher = spy(new ProcessApplicationEventPublisher(publisherMock));
     ApplicationReadyEvent applicationReadyEventMock = mock(ApplicationReadyEvent.class);
     processApplicationEventPublisher.handleApplicationReadyEvent(applicationReadyEventMock);
-    verify(processApplicationEventPublisher).publishProcessApplicationEvent(Mockito.any(ProcessApplicationStartedEvent.class));
+    verify(publisherMock).publishEvent(Mockito.any(ProcessApplicationStartedEvent.class));
   }
 
   @Test
   public void handleContextStoppedEventTest() {
-    ProcessApplicationEventPublisher processApplicationEventPublisher = spy(new ProcessApplicationEventPublisher(publisherMock));
     processApplicationEventPublisher.setApplicationContext(applicationContextMock);
     ContextClosedEvent contextClosedEventMock = mock(ContextClosedEvent.class);
     processApplicationEventPublisher.handleContextStoppedEvent(contextClosedEventMock);
-    verify(processApplicationEventPublisher, times(0)).publishProcessApplicationEvent(Mockito.any(ProcessApplicationStoppedEvent.class));
+
+    verify(publisherMock, never()).publishEvent(Mockito.any(ProcessApplicationStoppedEvent.class));
 
     when(contextClosedEventMock.getApplicationContext()).thenReturn(applicationContextMock);
     processApplicationEventPublisher.handleContextStoppedEvent(contextClosedEventMock);
-    verify(processApplicationEventPublisher).publishProcessApplicationEvent(Mockito.any(ProcessApplicationStoppedEvent.class));
+    verify(publisherMock).publishEvent(Mockito.any(ProcessApplicationStoppedEvent.class));
   }
 
 }
